Show stations on map using HERE api
Use HERE api as does not require billing and has decent documentation.
Most of index.js is standard code take from their documentation.

Possible to see the station details by clicking on the map marker. Went
for simple option of using standard marker, rather than anything
customised as a first option.

Kept the station list, in case user knows the name, but not the location
in the map.

Add a few decoder and type tests as an example.